requirements.txt
langchain
langchain-groq
langchain-community
chainlit
PyPDF2
chromadb
python-dotenv
nomic